# policy/api: use netip.Addr when sanitizing CIDR rules