This conversion should bring some memory usage improvements when using CIDR rules in policies.

Before/after benchmark:

```
name            old time/op    new time/op    delta
CIDRSanitize-8     969ns ± 7%     146ns ± 5%   -84.93%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
CIDRSanitize-8      168B ± 0%        0B       -100.00%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
CIDRSanitize-8      8.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```